update_gfwlist.sh

+++ entrypoint.sh
@@ -27,16 +27,19 @@ rm -rf wwwroot.tar.gz
 
 cat <<-EOF > /pac/update_gfwlist.sh
 #! /bin/bash
+if [[ -e /pac/gfwlist.txt.tmp ]]; then
+  return 1
+fi
+touch /pac/gfwlist.txt.tmp
 curl -o /pac/gfwlist.txt.tmp -L https://raw.githubusercontent.com/gfwlist/gfwlist/master/gfwlist.txt
 touch /pac/gfwlist.txt
 old_hash=\$(cat "/pac/gfwlist.txt" | sha256sum | awk '{print \$1}')
 new_hash=\$(cat "/pac/gfwlist.txt.tmp" | sha256sum | awk '{print \$1}')
 if [[ "\${old_hash}" == "\${new_hash}" ]]; then
-  rm -rf /pac/gfwlist.txt.tmp
   sed -i -r -e "s/(Check:).*/\1 \$(date '+%Y-%m-%d %H:%M:%S')/g" /pac/update.log
 else
   rm -rf /pac/cache/*
-  mv /pac/gfwlist.txt.tmp /pac/gfwlist.txt
+  cat /pac/gfwlist.txt.tmp > /pac/gfwlist.txt
   echo "/**" > /pac/update.log
   echo " * repository: https://github.com/ygcaicn/autopac-heroku" >> /pac/update.log
   echo " * /pac/gfwlist.txt Last-Modified: \$(date '+%Y-%m-%d %H:%M:%S')" >> /pac/update.log
@@ -44,7 +47,7 @@ else
   echo "*/" >> /pac/update.log
   echo "" >> /pac/update.log
 fi
-
+rm -rf /pac/gfwlist.txt.tmp
 EOF
 
 chmod +x /pac/update_gfwlist.sh
